v0.0.3.5

+++ src/args/posarg.rs
@@ -1,9 +1,7 @@
-
-#[derive(Clone)]
 pub struct PosArg {
     pub name: &'static str,
     pub help: Option<&'static str>,
     pub required: bool,
     pub value: Option<String>,
-    pub index: i32
+    pub index: u8 
 }
